+++ src/globals.c
@@ -16,24 +16,24 @@ hsize_t g_dimensions[2] = { 0, 0 };
 /**
  * Number of attributes
  */
-unsigned long g_n_attributes = 0;
+uint_fast32_t g_n_attributes = 0;
 
 /**
  * Number of observations
  */
-unsigned long g_n_observations = 0;
+uint_fast32_t g_n_observations = 0;
 
 /**
  * Number of classes
  */
-unsigned int g_n_classes = 0;
+uint_fast8_t g_n_classes = 0;
 
 /**
  * Number of bits used to store the class
  */
-unsigned int g_n_bits_for_class = 0;
+uint_fast8_t g_n_bits_for_class = 0;
 
 /**
  * Number of longs needed to store one line of the dataset
  */
-unsigned int g_n_longs = 0;
+uint_fast32_t g_n_longs = 0;
